Error on SegmentWriter re-use. Added guard against empty Or/And make_binary_tree error. Fixed delete_by_query to check for already-deleted documents.

+++ src/whoosh/writing.py
@@ -98,8 +98,9 @@ class IndexWriter(object):
         
         count = 0
         for docnum in q.docs(s):
-            self.delete_document(docnum)
-            count += 1
+            if not self.is_deleted(docnum):
+                self.delete_document(docnum)
+                count += 1
         
         if not searcher:
             s.close()
